Support Scala 3 final in all modules
