+++ dynamodb/src/main/scala-3/awscala/dynamodbv2/TableCompat.scala
@@ -1,0 +1,12 @@
+package awscala.dynamodbv2
+
+import DynamoDB.SimplePk
+
+private[dynamodbv2] trait TableCompat { self: Table =>
+  def putItem(hashPK: Any, attributes: SimplePk*)(implicit dynamoDB: DynamoDB): Unit = {
+    dynamoDB.put(this, hashPK, attributes: _*)
+  }
+  def putItem(hashPK: Any, rangePK: Any, attributes: SimplePk*)(implicit dynamoDB: DynamoDB): Unit = {
+    dynamoDB.put(this, hashPK, rangePK, attributes: _*)
+  }
+}
